Moved services to Services folder and namespace
