+++ OutlinesApp/OutlinesApp.csproj
@@ -60,9 +60,9 @@
     </ApplicationDefinition>
     <Compile Include="DimensionsConfig.cs" />
     <Compile Include="HoverWatcher.cs" />
-    <Compile Include="IColorPickerService.cs" />
-    <Compile Include="IScreenHelper.cs" />
-    <Compile Include="ServiceContainer.cs" />
+    <Compile Include="Services\IColorPickerService.cs" />
+    <Compile Include="Services\IScreenHelper.cs" />
+    <Compile Include="Services\ServiceContainer.cs" />
     <Compile Include="ViewModels\ColorPickerViewModel.cs" />
     <Compile Include="ViewModels\DimensionViewModel.cs" />
     <Compile Include="ViewModels\InspectorViewModel.cs" />
@@ -74,11 +74,11 @@
     <Compile Include="Views\OutlinesLayer.xaml.cs">
       <DependentUpon>OutlinesLayer.xaml</DependentUpon>
     </Compile>
-    <Compile Include="ScreenHelper.cs" />
+    <Compile Include="Services\ScreenHelper.cs" />
     <Compile Include="Views\PropertiesPanel.xaml.cs">
       <DependentUpon>PropertiesPanel.xaml</DependentUpon>
     </Compile>
-    <Compile Include="WindowHelper.cs" />
+    <Compile Include="Services\WindowHelper.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -88,7 +88,7 @@
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="ColorConfig.cs" />
-    <Compile Include="ColorPickerService.cs" />
+    <Compile Include="Services\ColorPickerService.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
