+++ OutlinesApp/Services/IColorPickerService.cs
@@ -1,7 +1,7 @@
 ﻿using System.Windows;
 using System.Windows.Media;
 
-namespace OutlinesApp
+namespace OutlinesApp.Services
 {
     public interface IColorPickerService
     {
