+++ OutlinesApp/ViewModels/DimensionViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using System.Windows;
 using Outlines;
+using OutlinesApp.Services;
 
 namespace OutlinesApp.ViewModels
 {
